Add files via upload
max() function can be used to find the global maximum number in an array. But what about a local max number? If we define a local max number as a number that is larger than the two numbers before and after it, i.e., x[n] > x[n-1] & x[n]>x[n+1]. This can be done in many different ways. In this code, I use diff and fliplr to find the max number.